[feat] modal detail 이미지뷰 바인딩 (#112)
