+++ Rollin_MVP/Rollin_MVP/Screens/DetailRollingPaper/DetailPostImageCollectionViewCell.swift
@@ -1,0 +1,37 @@
+//
+//  DetailPostImageCollectionViewCell.swift
+//  Rollin_MVP
+//
+//  Created by Seungyun Kim on 2022/11/15.
+//
+
+import UIKit
+
+final class DetailPostImageCollectionViewCell: UICollectionViewCell {
+    var imageView: UIImageView = {
+        let imageView = UIImageView()
+        imageView.layer.cornerRadius = 7
+        imageView.clipsToBounds = true
+        return imageView
+    }()
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setDetailPostLayout()
+    }
+    
+    private func setDetailPostLayout() {
+        self.contentView.addSubview(imageView)
+        imageView.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            imageView.leftAnchor.constraint(equalTo: contentView.leftAnchor),
+            imageView.rightAnchor.constraint(equalTo: contentView.rightAnchor),
+            imageView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor),
+            imageView.topAnchor.constraint(equalTo: contentView.topAnchor)
+        ])
+    }
+}
